Extract some types into Abstractions package

+++ code/benchmarks/Eshva.Caching.Nats.Benchmarks.Tools/InMemoryTestBuket.cs
@@ -27,7 +27,7 @@ public class InMemoryTestBuket : INatsObjStore, IBucketBackend {
 
   ValueTask<ObjectMetadata> INatsObjStore.PutAsync(string key, byte[] value, CancellationToken cancellationToken) {
     _entries[key] = value;
-    _metadata[key] = new ObjectMetadata { Name = key, Size = value.Length, Bucket = _bucket };
+    _metadata[key] = new ObjectMetadata { Name = key, Size = (ulong)value.Length, Bucket = _bucket };
     return ValueTask.FromResult(_metadata[key]);
   }
 
@@ -39,7 +39,7 @@ public class InMemoryTestBuket : INatsObjStore, IBucketBackend {
     using var reader = new BinaryReader(stream);
     _entries[key] = reader.ReadBytes((int)stream.Length);
     if (!leaveOpen) stream.Close();
-    _metadata[key] = new ObjectMetadata { Name = key, Size = (int)stream.Length, Bucket = _bucket };
+    _metadata[key] = new ObjectMetadata { Name = key, Size = (ulong)stream.Length, Bucket = _bucket };
     return ValueTask.FromResult(_metadata[key]);
   }
